implement validate payment amount endpoint in PaymentController
