+++ wallet/src/main/kotlin/net/thechance/wallet/service/WalletService.kt
@@ -1,15 +1,12 @@
 package net.thechance.wallet.service
 
 import net.thechance.wallet.repository.TransactionRepository
-import net.thechance.wallet.repository.WalletUserRepository
-import net.thechance.wallet.service.helper.PaymentAmountValidationResult
 import org.springframework.stereotype.Service
 import java.util.*
 
 @Service
 class WalletService(
     private val transactionRepository: TransactionRepository,
-    private val walletUserRepository: WalletUserRepository
 ) {
 
     fun getUserBalance(userId: UUID): Double {
@@ -19,19 +16,5 @@ class WalletService(
         val totalSent = transactionRepository.sumAmountBySenderId(userId) ?: 0.0
 
         return totalReceived - totalSent
-    }
-
-    fun validatePaymentAmount(userId: UUID, amount: Double, receiverId: UUID): PaymentAmountValidationResult {
-        val receiver = walletUserRepository.findById(receiverId)
-            .orElseThrow { IllegalArgumentException("Recipient not found") }
-
-        val currentBalance = getUserBalance(userId)
-        val isValid = amount > 0 && amount <= currentBalance
-        return PaymentAmountValidationResult(
-            isValid = isValid,
-            currentBalance = currentBalance,
-            receiverName = receiver.firstName + " " + receiver.lastName,
-            receiverImageUrl = receiver.imageUrl
-        )
     }
 }
